requirements.txt
autopep8==2.0.4
mathutils==3.3.0
numpy==1.25.2
Pillow==10.0.1
